requirements.txt
translitcodec
gunicorn
Flask
Flask-Security
Flask-SQLAlchemy
Flask-MySQL
